feat(workload): persist trainer training summaries in MongoDB

+++ workload-service/src/main/java/com/gymconnect/workload/dto/MonthSummaryResponse.java
@@ -8,6 +8,6 @@ package com.gymconnect.workload.dto;
  */
 public record MonthSummaryResponse(
         int month,
-        int trainingSummaryDuration
+        long trainingSummaryDuration
 ) {
 }
